Bruh
There was this weird issue where, no matter which replays folder I selected in the settings, it only wrote a slash
into the input field. Turns out, I thought the Qt directory chooser function returns a tuple, and as such I did `return_value[0]`.
But - no, it returns just a string. So I was simply chopping off the first character and returning that.